+++ Editor/BehaviourTreeView.cs
@@ -335,5 +335,15 @@ namespace TheKiwiCoder {
                 AddToSelection(nodeView);
             }
         }
+
+        public void UpdateEditorNodeSelectors(bool visibility)
+        {
+            foreach (var n in nodes)
+            {
+                NodeView nodeView = n as NodeView;
+                nodeView.UpdateDescriptionVisibility(visibility);
+                nodeView.UpdateErroredNode();
+            }
+        }
     }
 }