+++ Runtime/DecoratorNode.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 namespace TheKiwiCoder {
-    [BehaviourTreeNode(menuPath = "Decorator Nodes", nodeColor = NodeColors.blue)]
+    [BehaviourTreeNode(menuFolder = "Decorator Nodes", nodeColor = NodeColors.blue)]
     public abstract class DecoratorNode : Node {
 
         [SerializeReference]