+++ Runtime/Math/FloatDivideIntNode.cs
@@ -1,0 +1,7 @@
+﻿namespace TheKiwiCoder.Math
+{
+    public class FloatDivideIntNode
+    {
+        
+    }
+}