+++ Runtime/Math/FloatSubtractIntNode.cs
@@ -1,0 +1,7 @@
+﻿namespace TheKiwiCoder.Math
+{
+    public class FloatSubtractIntNode
+    {
+        
+    }
+}